aibeautyassistant/requirements.txt
TTS
numPy==1.21
openai
openai-whisper
gradio

